// legacy/protocol-sdk/src/apis/generated/premint-api-types.ts
export interface paths {
  "/signature": {
    /** Upsert Premint Signature */
    post: operations["upsert_premint_signature_signature_post"];
  };
  "/signature/{chain_name}/{collection_address}": {
    /** Get Premint Signatures For Collection */
    get: operations["get_premint_signatures_for_collection_signature__chain_name___collection_address__get"];
  };
  "/signature/{chain_name}/{collection_address}/next_uid": {
    /** Get Next Uid */
    get: operations["get_next_uid_signature__chain_name___collection_address__next_uid_get"];
  };
  "/signature/{chain_name}/{collection_address}/{uid}": {
    /** Get Premint Signature */
    get: operations["get_premint_signature_signature__chain_name___collection_address___uid__get"];
  };
  "/created/{chain_name}/{signer}": {
    /** Get Premint Signatures By Signer */
    get: operations["get_premint_signatures_by_signer_created__chain_name___signer__get"];
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    PremintRequest: {
      collection?: components["schemas"]["CollectionCreationConfig"];
      /** Premint */
      premint:
        | components["schemas"]["PremintConfigV1"]
        | components["schemas"]["PremintConfigV2"];
      chain_name: components["schemas"]["ChainName"];
      /** Signature */
      signature: string;
      /** Collection Address */
      collection_address?: string;
    };
// ...
    TokenCreationConfigV1: {
      /** Tokenuri */
      tokenURI: string;
      /** Maxsupply */
      maxSupply: string;
      /** Maxtokensperaddress */
      maxTokensPerAddress: string;
      /** Pricepertoken */
      pricePerToken: string;
      /** Mintstart */
      mintStart: string;
      /** Mintduration */
      mintDuration: string;
      /** Royaltymintschedule */
      royaltyMintSchedule: number;
      /** Royaltybps */
      royaltyBPS: number;
      /** Royaltyrecipient */
      royaltyRecipient: string;
      /** Fixedpriceminter */
      fixedPriceMinter: string;
    };
// ...
    TokenCreationConfigV2: {
      /** Tokenuri */
      tokenURI: string;
      /** Maxsupply */
      maxSupply: string;
      /** Maxtokensperaddress */
      maxTokensPerAddress: string;
      /** Pricepertoken */
      pricePerToken: string;
      /** Mintstart */
      mintStart: string;
      /** Mintduration */
      mintDuration: string;
      /** Royaltybps */
      royaltyBPS: number;
      /** Fixedpriceminter */
      fixedPriceMinter: string;
      /** Payoutrecipient */
      payoutRecipient: string;
      /** Createreferral */
      createReferral: string;
    };
    /** ValidationError */
    ValidationError: {
      /** Location */
      loc: (string | number)[];
      /** Message */
      msg: string;
      /** Error Type */
      type: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export interface operations {
  /** Upsert Premint Signature */
  upsert_premint_signature_signature_post: {
    requestBody: {
      content: {
        "application/json": components["schemas"]["PremintRequest"];
      };
    };
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          "application/json": components["schemas"]["PremintSignature"];
        };
      };
      /** @description Validation Error */
      422: {
        content: {
          "application/json": components["schemas"]["HTTPValidationError"];
        };
      };
    };
  };
  /** Get Premint Signatures For Collection */
  get_premint_signatures_for_collection_signature__chain_name___collection_address__get: {
    parameters: {
      path: {
        chain_name: components["schemas"]["ChainName"];
        collection_address: string;
      };
    };
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          "application/json": components["schemas"]["PremintCollection"];
        };
      };
      /** @description Validation Error */
      422: {
        content: {
          "application/json": components["schemas"]["HTTPValidationError"];
        };
      };
    };
  };
  /** Get Next Uid */
  get_next_uid_signature__chain_name___collection_address__next_uid_get: {
    parameters: {
      path: {
        chain_name: components["schemas"]["ChainName"];
        collection_address: string;
      };
    };
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          "application/json": components["schemas"]["PremintNextUid"];
        };
      };
      /** @description Validation Error */
      422: {
        content: {
          "application/json": components["schemas"]["HTTPValidationError"];
        };
      };
    };
  };
  /** Get Premint Signature */
  get_premint_signature_signature__chain_name___collection_address___uid__get: {
    parameters: {
      path: {
        chain_name: components["schemas"]["ChainName"];
        collection_address: string;
        uid: number;
      };
    };
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          "application/json": components["schemas"]["PremintSignature"];
        };
      };
      /** @description Validation Error */
      422: {
        content: {
          "application/json": components["schemas"]["HTTPValidationError"];
        };
      };
    };
  };
  /** Get Premint Signatures By Signer */
  get_premint_signatures_by_signer_created__chain_name___signer__get: {
    parameters: {
      path: {
        chain_name: components["schemas"]["ChainName"];
        signer: string;
      };
    };
    responses: {
      /** @description Successful Response */
      200: {
        content: {
          "application/json": components["schemas"]["PremintCollection"][];
        };
      };
      /** @description Validation Error */
      422: {
        content: {
          "application/json": components["schemas"]["HTTPValidationError"];
        };
      };
    };
  };
}